+++ packages/runtime/src/types/consumption/AttributeTagCollectionDTO.ts
