+++ packages/runtime/src/types/consumption/index.ts
@@ -1,3 +1,4 @@
+export * from "./AttributeTagCollectionDTO";
 export * from "./DraftDTO";
 export * from "./IdentityMetadataDTO";
 export * from "./LocalAttributeDTO";
@@ -6,4 +7,3 @@ export * from "./LocalNotificationDTO";
 export * from "./LocalRequestDTO";
 export * from "./RequestValidationResultDTO";
 export * from "./SettingDTO";
-export * from "./TagDTO";
